requirements.txt
streamlit>=1.28.0
PyPDF2>=3.0.0
python-docx>=1.1.0
google-generativeai>=0.3.0
